# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/area": {
      "post": {
        "description": "Permite crear una nueva especialidad o área médica (ej. Cardiología). \n    \nLas áreas agrupan a **Doctores** y **Servicios**. Es decir, antes de poder registrar un Doctor en Cardiología, debe existir el Área \"Cardiología\".\n\n**Roles permitidos**: ADMIN.",
        "operationId": "AreaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAreaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Área creada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cuid12345",
                    "name": "Cardiología",
                    "description": "Área especializada en el corazón",
                    "isActive": true,
                    "createdAt": "2026-06-03T10:00:00.000Z",
                    "updatedAt": "2026-06-03T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o el nombre del Área ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": [
                      "name must be a string"
                    ],
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Prohibido. Se requiere rol ADMIN.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una nueva área",
        "tags": [
          "Area"
        ]
      },
      "get": {
        "description": "Devuelve una lista con todas las áreas médicas del sistema. \n    \nÚtil para mostrar el menú de especialidades en el que un paciente puede buscar un Doctor o un Servicio.\nPuedes filtrar por áreas activas o inactivas enviando el parámetro opcional `?isActive=true` o `false`.\n\n**Roles permitidos**: ADMIN, DOCTOR, PATIENT.",
        "operationId": "AreaController_findAll",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de áreas obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "cuid12345",
                      "name": "Cardiología",
                      "description": "Área especializada en el corazón",
                      "isActive": true,
                      "createdAt": "2026-06-03T10:00:00.000Z",
                      "updatedAt": "2026-06-03T10:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Bad Request"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener todas las áreas",
        "tags": [
          "Area"
        ]
      }
    },
    "/area/{id}": {
      "get": {
        "description": "Devuelve la información detallada de una sola área buscando por su ID único.",
        "operationId": "AreaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Área obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cuid12345",
                    "name": "Cardiología",
                    "description": "Área especializada en el corazón",
                    "isActive": true,
                    "createdAt": "2026-06-03T10:00:00.000Z",
                    "updatedAt": "2026-06-03T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Área no encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Area not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener área por ID",
        "tags": [
          "Area"
        ]
      },
      "patch": {
        "description": "Permite modificar parcialmente los datos de un área existente (como su nombre o descripción). Solo disponible para usuarios con rol ADMIN.",
        "operationId": "AreaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAreaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Área actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cuid12345",
                    "name": "Cardiología Avanzada",
                    "description": "Área especializada en el corazón",
                    "isActive": true,
                    "createdAt": "2026-06-03T10:00:00.000Z",
                    "updatedAt": "2026-06-03T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (Bad Request).",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": [
                      "name must be a string"
                    ],
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Prohibido. Se requiere rol ADMIN.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Área no encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Area not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar área por ID",
        "tags": [
          "Area"
        ]
      }
    },
    "/area/{id}/disable": {
      "patch": {
        "description": "Cambia el estado de un área a inactiva en lugar de borrarla de la base de datos (Soft Delete). Solo disponible para usuarios con rol ADMIN.",
        "operationId": "AreaController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Área desactivada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cuid12345",
                    "name": "Cardiología",
                    "description": "Área especializada en el corazón",
                    "isActive": false,
                    "createdAt": "2026-06-03T10:00:00.000Z",
                    "updatedAt": "2026-06-03T11:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Prohibido. Se requiere rol ADMIN.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Área no encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Area not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar área por ID",
        "tags": [
          "Area"
        ]
      }
    },
...
  },
  "info": {
    "title": "API CLIMAS",
    "description": "Documentación API CLIMAS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAreaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del área",
            "example": "Cardiología"
          },
          "description": {
            "type": "string",
            "description": "Descripción del área",
            "example": "Especialidad centrada en el corazón"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del área",
            "example": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateAreaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del área",
            "example": "Cardiología"
          },
          "description": {
            "type": "string",
            "description": "Descripción del área",
            "example": "Especialidad centrada en el corazón"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del área",
            "example": true
          }
        }
      }
    }
  }
}